+++ src/app/Models/Models.php
@@ -65,4 +65,32 @@ class Models {
 	public function version():string {
 		return $this->PDO_ALL->query( 'select version()' )->fetchColumn();
 	}
+
+	public function doIExist( $tableName ) {
+		try {
+			$this->PDO_ALL->query( "SELECT 1 FROM {$tableName} LIMIT 1" );
+		} catch ( \Exception $e ) {
+			return false;
+		}
+
+		return true;
+	}
+
+	public function createSitelog() {
+		$sql = "CREATE TABLE IF NOT EXISTS `wpmgr_sitelog` (
+			`id` int(11) NOT NULL AUTO_INCREMENT,
+			`name` varchar(255) DEFAULT NULL,
+			`secure` tinyint(1) NOT NULL DEFAULT '0',
+			`emailreminder` tinyint(4) NOT NULL DEFAULT '0',
+			`created_by` varchar(45) DEFAULT NULL,
+			`created_date` datetime DEFAULT NULL,
+			`expiry_date` datetime DEFAULT NULL,
+			`deleted_by` varchar(45) DEFAULT NULL,
+			`deleted_date` datetime DEFAULT NULL,
+			PRIMARY KEY (`id`)
+		  ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb4;
+		  ";
+
+		$this->PDO_ALL->exec( $sql );
+	}
 }